Reduce compiler warnings
This code change started off to reduce the number of compiler warnings
building with this library produced, especially when used on a non AVR
based
systems such as the Teensy.  In my case a Teensy 4 and 4.1

The first change was to change all of the methods that had arguments
which passed strings as: char *
This causes warnings on systems if you pass in a string, such as all of
the helper functions do.
There was simple fix to convert these parameters to: const char *

There were also issues of using sscanf in the read functions, whenever
you had a conversion element like %d or %u and you did not pass
In a full size int.  My first fix for this was to convert these
variables to either: int or unsigned int.  But later I removed the usage
of sscanf

Likewise some warnings were due to things like passing "\r" to a
function that wants only a char.  Or likewise comparing a value to this
constant.
Which was easy fix.

But then went a bit farther and converted the code to not use things
like sscanf, but instead simply parse the data as it came in without
having
To first read into a buffer.  I also had the _s16 method simply call the
_str version and then parse the resulting string.